finally finished checking the microstates segmentation
