+++ neurokit2/microstates/__init__.py
@@ -6,7 +6,6 @@ from .microstates_static import microstates_static
 from .microstates_dynamic import microstates_dynamic
 from .microstates_complexity import microstates_complexity
 from .microstates_segment import microstates_segment
-from .microstates_quality import microstates_gev
 from .microstates_classify import microstates_classify
 from .microstates_plot import microstates_plot
 
@@ -17,6 +16,5 @@ __all__ = ["microstates_clean",
            "microstates_dynamic",
            "microstates_complexity",
            "microstates_segment",
-           "microstates_gev",
            "microstates_classify",
            "microstates_plot"]
